car_license_detection/requirements.txt
# pip install -r requirements.txt

matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.2
Pillow
PyYAML>=5.3.1
torch>=1.7.0
torchvision>=0.8.1
tqdm>=4.41.0
pandas
boto3
thop  # FLOPS computation
Werkzeug==0.15.4
Flask==1.1.2
Flask-RESTful==0.3.8
requests
python-dotenv
boto3